Warpped animation class
