+++ BlinderEngine/BlinderEngine.vcxproj
@@ -152,6 +152,7 @@
     <ClCompile Include="main.cpp" />
     <ClCompile Include="src\Camera.cpp" />
     <ClCompile Include="src\Cube.cpp" />
+    <ClCompile Include="src\DaeObject.cpp" />
     <ClCompile Include="src\graphic2D.cpp" />
     <ClCompile Include="src\Ground.cpp" />
     <ClCompile Include="src\Map.cpp" />
@@ -167,21 +168,22 @@
     <ClInclude Include="..\shared\GameData.h" />
     <ClInclude Include="..\shared\GlobalVariables.h" />
     <ClInclude Include="..\shared\Packet.h" />
-    <ClInclude Include="include\animation.h" />
-    <ClInclude Include="include\animator.h" />
-    <ClInclude Include="include\animdata.h" />
+    <ClInclude Include="include\Animation.h" />
+    <ClInclude Include="include\Animator.h" />
+    <ClInclude Include="include\Animdata.h" />
     <ClInclude Include="include\assimp_glm_helpers.h" />
-    <ClInclude Include="include\bone.h" />
+    <ClInclude Include="include\Bone.h" />
     <ClInclude Include="include\BShader.h" />
     <ClInclude Include="include\Camera.h" />
     <ClInclude Include="include\Constants.h" />
     <ClInclude Include="include\core.h" />
     <ClInclude Include="include\Cube.h" />
+    <ClInclude Include="include\DaeObject.h" />
     <ClInclude Include="include\graphic2D.h" />
     <ClInclude Include="include\Ground.h" />
     <ClInclude Include="include\Map.h" />
     <ClInclude Include="include\Mesh.h" />
-    <ClInclude Include="include\model_animation.h" />
+    <ClInclude Include="include\DynamicModel.h" />
     <ClInclude Include="include\StaticModel.h" />
     <ClInclude Include="include\MShader.h" />
     <ClInclude Include="include\ObjObject.h" />
