+++ BlinderEngine/BlinderEngine.vcxproj.filters
@@ -115,6 +115,9 @@
     <ClCompile Include="..\shared\GameData.cpp">
       <Filter>Shared</Filter>
     </ClCompile>
+    <ClCompile Include="src\DaeObject.cpp">
+      <Filter>src</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="include\Cube.h">
@@ -174,22 +177,28 @@
     <ClInclude Include="include\BShader.h">
       <Filter>include</Filter>
     </ClInclude>
-    <ClInclude Include="include\animdata.h">
+    <ClInclude Include="include\Animdata.h">
       <Filter>include</Filter>
     </ClInclude>
     <ClInclude Include="include\assimp_glm_helpers.h">
       <Filter>include</Filter>
     </ClInclude>
-    <ClInclude Include="include\bone.h">
+    <ClInclude Include="include\Bone.h">
       <Filter>include</Filter>
     </ClInclude>
-    <ClInclude Include="include\animation.h">
+    <ClInclude Include="include\Animation.h">
       <Filter>include</Filter>
     </ClInclude>
-    <ClInclude Include="include\animator.h">
+    <ClInclude Include="include\Animator.h">
       <Filter>include</Filter>
     </ClInclude>
-    <ClInclude Include="include\model_animation.h">
+    <ClInclude Include="include\DynamicModel.h">
+      <Filter>include</Filter>
+    </ClInclude>
+    <ClInclude Include="include\DaeObject.h">
+      <Filter>include</Filter>
+    </ClInclude>
+    <ClInclude Include="..\..\..\..\group-3\BlinderEngine\include\Timer.h">
       <Filter>include</Filter>
     </ClInclude>
     <ClInclude Include="..\client\Client.h">
