+++ BlinderEngine/include/Constants.h
@@ -2,6 +2,7 @@
 #define CONSTANTS_H
 
 #include <glm/glm.hpp>
+
 #include <string>
 
 namespace Constants
@@ -10,6 +11,9 @@ namespace Constants
 	const std::string nanosuit_object_path = "./resources/objects/nanosuit/nanosuit.obj";
 	const std::string backpack_object_path = "./resources/objects/backpack/backpack.obj";
 	const std::string rock_object_path = "./resources/objects/rock/rock.obj";
+	const std::string drone_object_path = "./resources/objects/uav/uav.dae";
+
+	const std::string drone_object_animation = "./resources/objects/uav/uav.dae";
 
 	// Define scaling factor for objects
 	const glm::vec3 nanosuit_scaling_factor = glm::vec3(0.5f, 0.5f, 0.5f);
